+++ apps/remix/app/routes/_h.$id.tsx
@@ -44,44 +44,41 @@ export default function MailDetail() {
   if (!mail) return <NotSelected></NotSelected>;
 
   return (
-    <>
-      <div>123</div>
-      <div className="flex flex-1 flex-col">
-        <div className="flex items-start p-4">
-          <div className="flex items-start gap-4 text-sm">
-            <Avatar>
-              <AvatarImage alt={mail.name} />
-              <AvatarFallback>
-                {mail.from.name
-                  .split(" ")
-                  .map((chunk) => chunk[0])
-                  .join("")}
-              </AvatarFallback>
-            </Avatar>
-            <div className="grid gap-1">
-              <div className="font-semibold">{mail.name}</div>
-              <div className="line-clamp-1 text-xs">{mail.subject}</div>
-              <div className="line-clamp-1 text-xs">
-                <span className="font-medium">Reply-To:</span>{" "}
-                {mail.to[0].address}
-              </div>
+    <div className="flex flex-1 flex-col">
+      <div className="flex items-start p-4">
+        <div className="flex items-start gap-4 text-sm">
+          <Avatar>
+            <AvatarImage alt={mail.name} />
+            <AvatarFallback>
+              {mail.from.name
+                .split(" ")
+                .map((chunk) => chunk[0])
+                .join("")}
+            </AvatarFallback>
+          </Avatar>
+          <div className="grid gap-1">
+            <div className="font-semibold">{mail.name}</div>
+            <div className="line-clamp-1 text-xs">{mail.subject}</div>
+            <div className="line-clamp-1 text-xs">
+              <span className="font-medium">Reply-To:</span>{" "}
+              {mail.to[0].address}
             </div>
           </div>
-          {mail.date && (
-            <div className="ml-auto text-xs text-muted-foreground">
-              {format(new Date(mail.date), "PPpp")}
-            </div>
-          )}
         </div>
-        <Separator />
-        <div
-          className="flex-1 whitespace-pre-wrap p-4 text-sm"
-          dangerouslySetInnerHTML={{
-            __html: mail.html || "",
-          }}
-        ></div>
-        <Separator className="mt-auto" />
+        {mail.date && (
+          <div className="ml-auto text-xs text-muted-foreground">
+            {format(new Date(mail.date), "PPpp")}
+          </div>
+        )}
       </div>
-    </>
+      <Separator />
+      <div
+        className="flex-1 whitespace-pre-wrap p-4 text-sm"
+        dangerouslySetInnerHTML={{
+          __html: mail.html || "",
+        }}
+      ></div>
+      <Separator className="mt-auto" />
+    </div>
   );
 }
